Fix "constant 2435016766 overflows int32" on GOARCH=arm
Fix #1

(cherry picked from commit dennwc/btrfs@ee67271a4cb46adc6cc1eab5eacb4561db6b9a54)

+++ utils.go
@@ -15,7 +15,7 @@ func isBtrfs(path string) (bool, error) {
 	if err := syscall.Statfs(path, &stfs); err != nil {
 		return false, &os.PathError{Op: "statfs", Path: path, Err: err}
 	}
-	return stfs.Type == SuperMagic, nil
+	return int64(stfs.Type) == SuperMagic, nil
 }
 
 func findMountRoot(path string) (string, error) {
